Move convert_circuit to a function
